+++ frontend/store.ts
@@ -1,34 +1,21 @@
 import { createStore, produce } from "solid-js/store";
 import { createResource } from "solid-js";
 import type { Message, UpdateMessage } from "../types/message";
+import type { Instance as InstanceData } from "../types/instance";
+import type { Info } from "../types/info";
 
-export type AppInfo = {
-  name: string;
-  iconUrl: string;
-  sourceCodeUrl: string;
-  manifestFound: boolean;
-  toolVersion: string;
-};
-
-const [appInfo] = createResource<AppInfo>(async () => {
+const [appInfo] = createResource<Info>(async () => {
   return await (await fetch("/app-info")).json();
 });
 
 export { appInfo };
 
-export type InstanceData = {
-  id: string;
-  url: string;
-  color: string;
-};
+const [instances, { refetch: refetchInstances, mutate: mutateInstances }] =
+  createResource<InstanceData[]>(async () => {
+    return (await fetch(`/instances`)).json();
+  });
 
-const [instances, { refetch: refetchInstances }] = createResource<
-  InstanceData[]
->(async () => {
-  return (await fetch(`/instances`)).json();
-});
-
-export { instances, refetchInstances };
+export { instances, refetchInstances, mutateInstances };
 
 const [state, setState] = createStore<Message[]>([]);